+++ vyper/simple/loop/array/simple_lower_half.vy
@@ -19,6 +19,6 @@ ARRAY_SIZE: constant(uint8) = 10
 @pure
 def main(array: uint8[ARRAY_SIZE]) -> uint8:
     sum: uint8 = 0
-    for i: uint256 in range(0, 5):
+    for i: uint8 in range(0, 5):
         sum += array[i]
     return sum
