+++ vyper/simple/loop/complex/1.vy
@@ -18,20 +18,20 @@ def test(param: uint8) -> uint256:
     cnt: uint256 = 0
     tmp: uint8 = 0
 
-    for i: uint256 in range(0, 2):
+    for i: uint8 in range(0, 2):
         if param == 1:
             return convert(param * 10, uint256) # 10
         if param == 21:
             break
         if param == 11:
             continue
-        for j: uint256 in [1, 2]:
+        for j: uint8 in [1, 2]:
             p: uint8 = 0
             if param == 22:
                 break
             if param == 33 and j > 1:
                 return cnt
-            for _: uint256 in range(1000000):
+            for _: uint8 in range(1000000):
                 if param == 3:
                     return convert(param * 10, uint256) # 30
                 p += 2
@@ -41,14 +41,14 @@ def test(param: uint8) -> uint256:
                     continue
                 if param == 23:
                     break
-                for h: uint256 in range(1, 5):
+                for h: uint8 in range(1, 5):
                     if param == 4:
                         return 40
                     if h > 2:
                         break
                     if param == 24:
                         break
-                    for k: uint256 in range(10, 12):
+                    for k: uint8 in range(10, 12):
                         x: uint8 = 7
                         tmp = 5 * k
                         if param == 5:
@@ -80,7 +80,7 @@ def test(param: uint8) -> uint256:
                                     continue
                                 if param == 32 and y > 13:
                                     break
-                                for l: uint256 in range(4):
+                                for l: uint8 in range(4):
                                     if param == 28:
                                         break
                                     if l % 2 == 0:
