+++ vyper/simple/loop/complex/3.vy
@@ -15,17 +15,17 @@
 @internal
 @pure
 def test() -> uint256:
-    for i: uint256 in range(0, 2):
-        for j: uint256 in [1, 2]:
+    for i: uint8 in range(0, 2):
+        for j: uint8 in [1, 2]:
             p: uint8 = 0
             for _: uint256 in range(1000000):
                 p += 2
                 if p == 8:
                     break
-                for h: uint256 in range(1, 5):
+                for h: uint8 in range(1, 5):
                     if h > 2:
                         break
-                    for k: uint256 in range(10, 12):
+                    for k: uint8 in range(10, 12):
                         x: uint8 = 6
                         for __: uint256 in range(1000000):
                             x -= 1
