+++ vyper/simple/loop/range_bound/i16.vy
@@ -16,6 +16,6 @@
 @pure
 def main() -> uint64:
     sum: uint64 = 0
-    for i: uint256 in range(-32768, -32758):
+    for i: int16 in range(-32768, -32758):
         sum += 1
     return sum
