+++ vyper/simple/loop/range_bound/i8.vy
@@ -16,6 +16,6 @@
 @pure
 def main() -> uint64:
     sum: uint64 = 0
-    for i: uint256 in range(-128, -118):
+    for i: int8 in range(-128, -118):
         sum += 1
     return sum
